장바구니 아이템 클래스 interface 폴더에 추가
